Rename `ai` to `interpreter`
