+++ crates/profiler-lib/src/interpreter.rs
